pyproject.toml
[tool.black]
line-length = 120
skip-string-normalization = '1'
exclude = '''
/(
  | env
  | proto
  |local_settings.py
)/
'''


[tool.poetry]
name = "fastapi-project-template"
version = "3.15.4"
description = ""
authors = ["Anton Ogorodnikov"]

[tool.poetry.dependencies]
"ruamel.yaml" = "^0.16.10"
click = "7.*"
fastapi = "0.*"
orjson = "3.*"
poetry = "1.1.4"
pydantic = "~1.5"
python = "3.9.0"
python-multipart = "0.*" # for fastapi
uvicorn = "0.*"

[tool.poetry.dev-dependencies]
asynctest = "0.*"
autoflake = "1.*"
black = "19.10b0"
flake8 = "3.*"
isort = "4.*"
lxml = "^4.6.2"
mypy = "0.*"
pip = "20.*"
pytest = "6.*"
pytest-asyncio = "0.*"
pytest-cov = "2.*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
